Re-merge AD-dependent vector checks into VectorBijectors.test_all
Restore DifferentiationInterface as a hard Bijectors dependency and fold
test_linked_optic, test_linked_logjac, test_ad, the
to_vec_for_logjac_test/from_vec_for_logjac_test family, and _rand_safe_ad
back into src/vector/test_utils.jl. test_all gains adtypes / ad_atol /
ad_rtol kwargs and defaults adtypes to [AutoForwardDiff()] so the main
Vector suite exercises FD as the baked-in reference. Integration suites
pass their own adtypes list through the (now overloaded) run_vector_case.
